+++ packages/react-next/src/components/OverflowSet/OverflowButton.tsx
@@ -1,0 +1,5 @@
+import { getStyles } from './OverflowSet.styles';
+import { styled } from '../../Utilities';
+import { OverflowButtonBase } from './OverflowButton.base';
+
+export const OverflowButton = styled(OverflowButtonBase, getStyles, undefined, { scope: 'OverflowButton' });