+++ packages/react-next/src/components/OverflowSet/index.ts
@@ -1,3 +1,4 @@
 export * from './OverflowSet';
 export * from './OverflowSet.base';
 export * from './OverflowSet.types';
+export * from './OverflowButton.base';